dev-requirements.txt
python-dotenv==1.0.0
coverage==7.8.0
